+++ libc/calls/execve-sysv.c
@@ -25,7 +25,6 @@
 #include "libc/errno.h"
 #include "libc/intrin/bits.h"
 #include "libc/intrin/safemacros.internal.h"
-#include "libc/intrin/weaken.h"
 #include "libc/mem/alloca.h"
 #include "libc/paths.h"
 #include "libc/runtime/runtime.h"
@@ -34,7 +33,6 @@
 #include "libc/sysv/consts/o.h"
 #include "libc/sysv/consts/ok.h"
 #include "libc/sysv/errfuns.h"
-#include "libc/zipos/zipos.internal.h"
 
 static bool CanExecute(const char *path) {
   return !sys_faccessat(AT_FDCWD, path, X_OK, 0);
@@ -77,19 +75,6 @@ int sys_execve(const char *prog, char *const argv[], char *const envp[]) {
   char *buf;
   char **shargs;
   const char *ape;
-  struct ZiposUri uri;
-  if (_weaken(__zipos_parseuri) &&
-      (_weaken(__zipos_parseuri)(prog, &uri) != -1)) {
-    rc = _weaken(__zipos_open)(&uri, O_RDONLY | O_CLOEXEC, 0);
-    if (rc != -1) {
-      const int zipFD = rc;
-      strace_enabled(-1);
-      rc = fexecve(zipFD, argv, envp);
-      close(zipFD);
-      strace_enabled(+1);
-    }
-    return rc;
-  }
   e = errno;
   __sys_execve(prog, argv, envp);
   if (errno == ENOEXEC) {
